Add `nginxingress` component to Seed reconciliation flow
