+++ pkg/operation/botanist/component/nginxingress/nginxingress.go
@@ -75,11 +75,6 @@ var (
 	ingressClassName = v1beta1constants.SeedNginxIngressClass122
 )
 
-// Interface contains functions for a nginxIngress Deployer
-type Interface interface {
-	component.DeployWaiter
-}
-
 // Values is a set of configuration values for the nginxIngress component.
 type Values struct {
 	// ImageController is the container image used for nginxIngress Controller.
@@ -95,7 +90,7 @@ func New(
 	client client.Client,
 	namespace string,
 	values Values,
-) Interface {
+) component.DeployWaiter {
 	return &nginxIngress{
 		client:    client,
 		namespace: namespace,
